完成了除读写文件具体内容外的所有任务，暂未测试

+++ include/fs/bitmap.h
@@ -2,7 +2,7 @@
 #define __BITMAP_H__
 
 uint32 bitmap_alloc_block();
-uint32 bitmap_alloc_inode();
+uint16 bitmap_alloc_inode();
 void   bitmap_free_block(uint32 block_num);
 void   bitmap_free_inode(uint16 inode_num);
 void   bitmap_print(uint32 bitmap_block_num);
